+++ src/global_functions.php
@@ -13,6 +13,9 @@ declare(strict_types=1);
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 
+use Symfony\Component\String\Slugger\AsciiSlugger;
+use Tools\Exceptionist;
+
 if (!defined('IS_WIN')) {
     define('IS_WIN', DIRECTORY_SEPARATOR === '\\');
 }
@@ -149,7 +152,7 @@ if (!function_exists('deprecationWarning')) {
      * @return void
      * @since 1.1.7
      */
-    function deprecationWarning(string $message, int $stackFrame = 1): void
+    function deprecationWarning(string $message, int $stackFrame = 0): void
     {
         if (!(error_reporting() & E_USER_DEPRECATED)) {
             return;
@@ -275,7 +278,7 @@ if (!function_exists('objects_map')) {
     function objects_map(array $objects, string $method, array $args = []): array
     {
         return array_map(function ($object) use ($method, $args) {
-            is_true_or_fail(method_exists($object, '__call') || method_exists($object, $method), sprintf(
+            Exceptionist::isTrue(method_exists($object, '__call') || method_exists($object, $method), sprintf(
                 'Class `%s` does not have a method `%s`',
                 get_class($object),
                 $method
@@ -283,6 +286,23 @@ if (!function_exists('objects_map')) {
 
             return call_user_func_array([$object, $method], $args);
         }, $objects);
+    }
+}
+
+if (!function_exists('slug')) {
+    /**
+     * Gets a slug from a string
+     * @param string $string The string you want to generate the slug from
+     * @param bool $lowerCase With `true` the string will be lowercase
+     * @return string
+     * @see https://symfony.com/doc/current/components/string.html#slugger
+     * @since 1.4.1
+     */
+    function slug(string $string, bool $lowerCase = true): string
+    {
+        $slug = (string)(new AsciiSlugger())->slug($string);
+
+        return $lowerCase ? strtolower($slug) : $slug;
     }
 }